Add SimpleRepository
