+++ autogluon_zeroshot/repository/__init__.py
@@ -1,0 +1,2 @@
+from .repository import SimpleRepository
+from .evaluation_repository import EvaluationRepository
