+++ autogluon_zeroshot/repository/repository.py
@@ -1,26 +1,13 @@
-from pathlib import Path
-from typing import Optional, Union
-
-import pandas as pd
-import numpy as np
-
-from autogluon.common.loaders import load_pkl
-from autogluon.common.savers import save_pkl
+from typing import Union
 
 from typing import List
-from autogluon_zeroshot.simulation.configuration_list_scorer import ConfigurationListScorer
-from autogluon_zeroshot.simulation.ensemble_selection_config_scorer import EnsembleSelectionConfigScorer
-from autogluon_zeroshot.simulation.single_best_config_scorer import SingleBestConfigScorer
 from autogluon_zeroshot.simulation.simulation_context import ZeroshotSimulatorContext
 from autogluon_zeroshot.simulation.tabular_predictions import TabularModelPredictions
+from autogluon_zeroshot.utils.cache import SaveLoadMixin
 from autogluon_zeroshot.utils import catchtime
-from autogluon_zeroshot.contexts.utils import intersect_folds_and_datasets, force_to_dense, prune_zeroshot_gt
-from autogluon_zeroshot.portfolio import Portfolio, PortfolioCV
-from autogluon_zeroshot.simulation.sim_runner import run_zs_simulation
-from autogluon_zeroshot.simulation.sim_output import SimulationOutputGenerator
 
 
-class EvaluationRepository:
+class SimpleRepository(SaveLoadMixin):
     def __init__(
             self,
             zeroshot_context: ZeroshotSimulatorContext,
@@ -43,19 +30,6 @@ class EvaluationRepository:
     def _tid_to_name(self):
         return {v: k for k, v in self._name_to_tid.items()}
 
-    def save(self, path: Union[str, Path]):
-        path = str(path)
-        assert path.endswith('.pkl')
-        save_pkl.save(path=path, object=self)
-
-    @classmethod
-    def load(cls, path: Union[str, Path]):
-        path = str(path)
-        assert path.endswith('.pkl')
-        obj = load_pkl.load(path=path)
-        assert isinstance(obj, cls)
-        return obj
-
     def subset(self,
                folds: List[int] = None,
                models: List[str] = None,
@@ -75,6 +49,8 @@ class EvaluationRepository:
         return self
 
     def force_to_dense(self, verbose: bool = True):
+        # TODO: Move these util functions to simulations or somewhere else to avoid circular imports
+        from autogluon_zeroshot.contexts.utils import intersect_folds_and_datasets, force_to_dense, prune_zeroshot_gt
         # keep only dataset whose folds are all present
         intersect_folds_and_datasets(self._zeroshot_context, self._tabular_predictions, self._ground_truth)
         force_to_dense(self._tabular_predictions,
@@ -158,154 +134,22 @@ class EvaluationRepository:
         metadata = self._df_metadata[self._df_metadata.name == dataset_name]
         return dict(zip(metadata.columns, metadata.values[0]))
 
-    def evaluate_ensemble(
-            self,
-            dataset_names: List[str],
-            config_names: List[str],
-            ensemble_size: int,
-            rank: bool = True,
-            folds: Optional[List[int]] = None,
-            backend: str = "ray",
-    ) -> np.array:
-        """
-        :param dataset_names: list of dataset to compute errors on.
-        :param config_names: list of config to consider for ensembling.
-        :param ensemble_size: number of members to select with Caruana.
-        :param rank: whether to return ranks or raw scores (e.g. RMSE). Ranks are computed over all base models and
-        automl framework.
-        :param folds: list of folds that need to be evaluated, use all folds if not provided.
-        :return: 2D array of scores whose rows are datasets and columns are folds
-        """
-        if folds is None:
-            folds = range(self.n_folds())
-        dataset_fold_name = lambda dataset, fold: f"{self.dataset_to_taskid(dataset)}_{fold}"
-        tasks = [
-            dataset_fold_name(dataset, fold)
-            for dataset in dataset_names
-            for fold in folds
-        ]
-        scorer = self._construct_ensemble_selection_config_scorer(
-            datasets=tasks,
-            ensemble_size=ensemble_size,
-            backend=backend,
-        )
-        if rank:
-            dict_scores = scorer.score_per_dataset(config_names)
-        else:
-            dict_scores = scorer.compute_errors(configs=config_names)
-
-        return np.array([[
-                dict_scores[dataset_fold_name(dataset, fold)
-            ] for fold in folds
-        ] for dataset in dataset_names])
-
     def get_datasets(self, problem_type=None):
         return self._zeroshot_context.get_datasets(problem_type=problem_type)
 
     def n_folds(self) -> int:
         return len(self._tabular_predictions.folds)
 
-    def _construct_config_scorer(self,
-                                 config_scorer_type: str = 'ensemble',
-                                 **config_scorer_kwargs) -> ConfigurationListScorer:
-        if config_scorer_type == 'ensemble':
-            return self._construct_ensemble_selection_config_scorer(**config_scorer_kwargs)
-        elif config_scorer_type == 'single':
-            return self._construct_single_best_config_scorer(**config_scorer_kwargs)
-        else:
-            raise ValueError(f'Invalid config_scorer_type: {config_scorer_type}')
 
-    def _construct_ensemble_selection_config_scorer(self,
-                                                    ensemble_size: int = 10,
-                                                    backend='ray',
-                                                    **kwargs) -> EnsembleSelectionConfigScorer:
-        config_scorer = EnsembleSelectionConfigScorer.from_zsc(
-            zeroshot_simulator_context=self._zeroshot_context,
-            zeroshot_gt=self._ground_truth,
-            zeroshot_pred_proba=self._tabular_predictions,
-            ensemble_size=ensemble_size,  # 100 is better, but 10 allows to simulate 10x faster
-            backend=backend,
-            **kwargs,
-        )
-        return config_scorer
-
-    def _construct_single_best_config_scorer(self, **kwargs) -> SingleBestConfigScorer:
-        config_scorer = SingleBestConfigScorer.from_zsc(
-            zeroshot_simulator_context=self._zeroshot_context,
-            **kwargs,
-        )
-        return config_scorer
-
-    # TODO: add saving support
-    # TODO: add loading support
-    def generate_output_from_portfolio_cv(self,
-                                          portfolio_cv: PortfolioCV,
-                                          name: str,
-                                          config_scorer_type: str = 'ensemble',
-                                          config_scorer_kwargs: dict = None) -> pd.DataFrame:
-        sog = SimulationOutputGenerator(repo=self,
-                                        config_scorer_type=config_scorer_type,
-                                        config_scorer_kwargs=config_scorer_kwargs)
-        df_result = sog.from_portfolio_cv(portfolio_cv=portfolio_cv, name=name)
-        return df_result
-
-    # TODO: add saving support
-    # TODO: add loading support
-    def generate_output_from_portfolio(self,
-                                       portfolio: Union[Portfolio, List[str]],
-                                       name: str,
-                                       config_scorer_type: str = 'ensemble',
-                                       config_scorer_kwargs: dict = None) -> pd.DataFrame:
-        sog = SimulationOutputGenerator(repo=self,
-                                        config_scorer_type=config_scorer_type,
-                                        config_scorer_kwargs=config_scorer_kwargs)
-        df_result = sog.from_portfolio(portfolio=portfolio, name=name)
-        return df_result
-
-    # TODO: add simulate_zeroshot_debug
-    # TODO: add saving support
-    # TODO: add loading support
-    def simulate_zeroshot(self,
-                          num_zeroshot: int = 10,
-                          n_splits: int = 2,
-                          backend: str = 'ray',
-                          config_scorer_type: str = 'ensemble',
-                          config_scorer_kwargs: dict = None) -> PortfolioCV:
-        if config_scorer_kwargs is None:
-            config_scorer_kwargs = {}
-        if config_scorer_type == 'ensemble':
-            config_scorer = self._construct_ensemble_selection_config_scorer(**config_scorer_kwargs)
-        elif config_scorer_type == 'single':
-            config_scorer = self._construct_single_best_config_scorer(**config_scorer_kwargs)
-        else:
-            raise ValueError(f'Unknown config_scorer_type: {config_scorer_type}')
-        results_cv: PortfolioCV = run_zs_simulation(
-            zsc=self._zeroshot_context,
-            config_scorer=config_scorer,
-            n_splits=n_splits,
-            config_generator_kwargs={'num_zeroshot': num_zeroshot},
-            backend=backend,
-        )
-        return results_cv
-
-
-def load(version: str = None):
-    from autogluon_zeroshot.contexts import get_context
-    zsc, configs_full, zeroshot_pred_proba, zeroshot_gt = get_context(version).load(load_predictions=True, lazy_format=False)
-    return EvaluationRepository(
-        zeroshot_context=zsc,
-        tabular_predictions=zeroshot_pred_proba,
-        ground_truth=zeroshot_gt,
-    )
-
-
+# TODO: git shelve ADD BACK
 if __name__ == '__main__':
+    from autogluon_zeroshot.contexts.context_artificial import load_context_artificial
+    from autogluon_zeroshot.repository.evaluation_repository import EvaluationRepository
     with catchtime("loading repo and evaluating one ensemble config"):
         dataset_name = "abalone"
         config_name = "NeuralNetFastAI_r1"
         # repo = EvaluationRepository.load(version="2022_10_13")
 
-        from autogluon_zeroshot.contexts.context_artificial import load_context_artificial
         zsc, configs_full, zeroshot_pred_proba, zeroshot_gt = load_context_artificial()
         repo = EvaluationRepository(
             zeroshot_context=zsc,
